# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exam Backend API",
    "description": "API documentation for Sindre Gangeskar's Exam Project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "adminAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    },
    "userAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Include **Bearer** prefix in the value field with token - e.g **Bearer abc123**"
    }
  },
  "paths": {
...
    "/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "A **logged in user** can retrieve all items that have been added to the user\\'s cart. If the cart is empty, an **empty array** will be returned",
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully retrieves all items placed in cart for current user, if no cart items are found an **empty array** is returned",
            "schema": {
              "$ref": "#/definitions/cartGetSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "500": {
            "description": "This response is returned when **an internal server** has occurred while the API is trying to get all items in cart for the current user",
            "schema": {
              "$ref": "#/definitions/cartGetInternalError"
            }
          }
        },
        "security": [
          {
            "userAuth": []
          },
          {
            "adminAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "A **logged in user** can add a product to the cart",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "The **product id** must be provided in order to add a product to cart",
            "schema": {
              "type": "object",
              "properties": {
                "ProductId": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully adds a product to the cart",
            "schema": {
              "$ref": "#/definitions/cartPostSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the **user** provides incorrect or invalid values or properties in the body",
            "schema": {
              "$ref": "#/definitions/cartPostUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API is unable to find a product with the provided id the user wishes to add to cart",
            "schema": {
              "$ref": "#/definitions/cartPostNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to add a product to cart",
            "schema": {
              "$ref": "#/definitions/cartPostInternalError"
            }
          }
        },
        "security": [
          {
            "userAuth": []
          },
          {
            "adminAuth": []
          }
        ]
      }
    },
    "/cart/checkout/now": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "A **logged in user** can proceed to checkout and an **order** will be created with a unique **order number**",
        "responses": {
          "200": {
            "description": "This response is returned when the API succeeds in checking out all cart products in the cart for the logged in user",
            "schema": {
              "$ref": "#/definitions/cartCheckoutSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the API fails to find items in the cart for the logged in user",
            "schema": {
              "$ref": "#/definitions/cartCheckoutUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "500": {
            "description": "This response is returned when **an internal error** occurrs while the API is trying to check out all of the items placed in the cart for the logged in user",
            "schema": {
              "$ref": "#/definitions/cartCheckoutInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          },
          {
            "userAuth": []
          }
        ]
      }
    },
    "/cart/{id}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "A **logged in user** can remove an item from its cart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The cart item id"
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully deleted a cart item",
            "schema": {
              "$ref": "#/definitions/cartDeleteSuccess"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API is unable to find a cart item with the id provided by the user",
            "schema": {
              "$ref": "#/definitions/cartDeleteNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **an internal error** occurrs while the API is trying remove an item from cart",
            "schema": {
              "$ref": "#/definitions/cartDeleteInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          },
          {
            "userAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Cart"
        ],
        "description": "A **logged in user** can update the item\\'s quantity by using **Add**, **Remove** or **TotalQuantity** in the body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The cart item id"
          },
          {
            "name": "body",
            "in": "body",
            "description": "The properties available. Only **one** property is allowed at a time",
            "schema": {
              "type": "object",
              "properties": {
                "totalQuantity": {
                  "type": "number",
                  "example": 1
                },
                "add": {
                  "type": "number",
                  "example": 10
                },
                "remove": {
                  "type": "number",
                  "example": 5
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "This response is returned when the API successfully updates a cart item",
            "schema": {
              "$ref": "#/definitions/cartUpdateSuccess"
            }
          },
          "400": {
            "description": "This response is returned when the **user** has provided incorrect or invalid values or properties in the body",
            "schema": {
              "$ref": "#/definitions/cartUpdateUserError"
            }
          },
          "401": {
            "description": "This response is returned when the user isn't logged in. The response will also be returned if the **session has expired** or the **token is invalid or malformed**",
            "schema": {
              "$ref": "#/definitions/unauthorized"
            }
          },
          "404": {
            "description": "This response is returned when the API is unable to find a cart item by the id the **user** provided",
            "schema": {
              "$ref": "#/definitions/cartUpdateNotFound"
            }
          },
          "500": {
            "description": "This response is returned when an **internal server error** occurrs while the API is trying to update a cart item",
            "schema": {
              "$ref": "#/definitions/cartUpdateInternalError"
            }
          }
        },
        "security": [
          {
            "adminAuth": []
          },
          {
            "userAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "unauthorized": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 401
            },
            "result": {
              "type": "array",
              "example": [
                "Login required",
                "Invalid token, please log in again",
                "Session has expired, please log in again"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidJWTError",
                "InvalidCredentialsError",
                "UnauthorizedError",
                "TokenExpiredError"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "cartPostSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully added product to cart"
            }
          }
        }
      }
    },
    "cartPostNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "array",
              "example": [
                "Cannot add product to cart - Product with provided ProductId does not exist"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "string",
              "example": "CartProductNotFoundError"
            }
          }
        }
      }
    },
    "cartPostInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to add product to cart"
            }
          }
        }
      }
    },
    "cartPostUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "array",
              "example": [
                "Required properties are missing",
                "Please remove invalid properties",
                "ProductId must be a number",
                "Cannot add product to cart - product is no longer available"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "InvalidPropertyError",
                "CartProductUnavailableError",
                "InvalidNumberError"
              ],
              "items": {
                "type": "string"
              }
            },
            "missingInBody": {
              "type": "array",
              "example": [
                "ProductId"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "InvalidPropertyExample1",
                "InvalidPropertyExample2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "cartCheckoutSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully purchased items"
            },
            "ordernumber": {
              "type": "string",
              "example": "abc123ef"
            }
          }
        }
      }
    },
    "cartCheckoutUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "string",
              "example": "Cannot process order - No items in cart"
            },
            "error": {
              "type": "string",
              "example": "NoItemsInCartError"
            }
          }
        }
      }
    },
    "cartCheckoutInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to create order"
            }
          }
        }
      }
    },
    "cartGetSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "array",
              "example": [
                "Successfully found cart item(s)",
                "Cart is empty"
              ],
              "items": {
                "type": "string"
              }
            },
            "cart": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "product": {
                    "type": "string",
                    "example": "Mac Pro Quad-Core and Dual Core GPU"
                  },
                  "ProductId": {
                    "type": "number",
                    "example": 8
                  },
                  "imgurl": {
                    "type": "string",
                    "example": "http://images.restapi.co.za/products/product-mac-pro.png"
                  },
                  "quantity": {
                    "type": "number",
                    "example": 2
                  },
                  "unitprice": {
                    "type": "number",
                    "example": 2999
                  },
                  "totalPrice": {
                    "type": "number",
                    "example": 5998
                  },
                  "membership": {
                    "type": "string",
                    "example": "Bronze"
                  },
                  "MembershipId": {
                    "type": "number",
                    "example": 1
                  },
                  "discount": {
                    "type": "number",
                    "example": 0
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2024-11-06T21:40:39.000Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2024-11-06T21:40:50.000Z"
                  }
                }
              }
            }
          }
        }
      }
    },
    "cartGetInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to get all cart records"
            }
          }
        }
      }
    },
...
    "cartDeleteSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully deleted item from cart"
            }
          }
        }
      }
    },
    "cartDeleteNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statuaCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "string",
              "example": "No item with provided cart id exists"
            },
            "error": {
              "type": "string",
              "example": "CartItemNotFoundError"
            }
          }
        }
      }
    },
    "cartDeleteInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to delete an item from the cart"
            }
          }
        }
      }
    },
    "cartUpdateSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 200
            },
            "result": {
              "type": "string",
              "example": "Successfully updated item in cart"
            }
          }
        }
      }
    },
    "cartUpdateNotFound": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 404
            },
            "result": {
              "type": "string",
              "example": "Cannot update cart - Cannot find cart item with provided id"
            },
            "error": {
              "type": "string",
              "example": "CartItemNotFoundError"
            }
          }
        }
      }
    },
    "cartUpdateUserError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 400
            },
            "result": {
              "type": "array",
              "example": [
                "At least one property must be present in the body",
                "Please remove invalid properties",
                "Cannot set total quantity - totalQuantity must be 1 or higher",
                "Cannot add to cart - add must be 1 or higher",
                "Cannot add to cart - Specified amount to add exceeds the available product quantity",
                "Cannot remove from cart - remove must be 1 or higher",
                "Cannot remove any more from cart - At least 1 in quantity is required in cart. Consider deleting the cart if you want to remove all",
                "Only one property can be used at a time - add, remove or totalQuantity",
                "totalQuantity must be a number",
                "add must be a number",
                "remove must be a number"
              ],
              "items": {
                "type": "string"
              }
            },
            "error": {
              "type": "array",
              "example": [
                "CartUpdateRemoveError",
                "CartUpdateError",
                "CartUpdateAddError",
                "CartUpdateTotalQuantityError",
                "InvalidNumberError"
              ],
              "items": {
                "type": "string"
              }
            },
            "validProperties": {
              "type": "array",
              "example": [
                "totalQuantity",
                "add",
                "remove"
              ],
              "items": {
                "type": "string"
              }
            },
            "invalidProperties": {
              "type": "array",
              "example": [
                "InvalidPropertyExample1",
                "InvalidPropertyExample2"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "cartUpdateInternalError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "error"
        },
        "data": {
          "type": "object",
          "properties": {
            "statusCode": {
              "type": "number",
              "example": 500
            },
            "result": {
              "type": "string",
              "example": "An internal server error has occurred while trying to update cart item"
            }
          }
        }
      }
    },
...
  }
}